// src/sdks/systemV2ApiDocs.ts
export type getSystemDicttypePageProps = {
  dictName?: string | null | undefined;
  dictType?: string | null | undefined;
  pageNum: number;
  pageSize: number;
};
// ...
export interface DictTypeResp {
  dictName?: string;

  dictType?: string;

  id?: number;
}